src/ml/requirements.txt
# ML Requirements for CQC Rating Predictor

# Google Cloud AI Platform
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.13.0
google-cloud-storage==2.13.0
google-cloud-logging==3.8.0

# Vertex AI Extensions
google-cloud-aiplatform[prediction]==1.38.1
google-cloud-aiplatform[tensorboard]==1.38.1
google-cloud-aiplatform[training]==1.38.1

# Machine Learning Frameworks
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
tensorflow==2.15.0
tensorflow-io==0.35.0

# ML Utilities
imbalanced-learn==0.11.0
category_encoders==2.6.3
feature-engine==1.6.2

# Model Interpretability
shap==0.43.0
lime==0.2.0.1

# Hyperparameter Optimization
optuna==3.5.0
hyperopt==0.2.7

# Experiment Tracking
mlflow==2.9.2
tensorboard==2.15.1
wandb==0.16.1

# Data Processing
pandas==2.1.4
numpy==1.24.4
pyarrow==14.0.2
dask[complete]==2023.12.1

# Feature Engineering
featuretools==1.28.0
tsfresh==0.20.1

# Model Evaluation
scikit-plot==0.3.7
yellowbrick==1.5

# Statistical Analysis
scipy==1.11.4
statsmodels==0.14.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
hydra-core==1.3.2

# Utilities
joblib==1.3.2
tqdm==4.66.1
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.7.1
pylint==3.0.3

# Jupyter Development
jupyter==1.0.0
jupyterlab==4.0.9
ipywidgets==8.1.1

# Type Hints
pandas-stubs==2.1.4.231227
types-requests==2.31.0.10